New webroot file

+++ webroot/wp-content/plugins/ibiza-categories-pull-plugin/IbizaCategoriesPullPlugin_Plugin.php
@@ -1,9 +1,9 @@
 <?php
-
 include_once('IbizaCategoriesPullPlugin_LifeCycle.php');
 
 define('TOP_NODE', 'categories');
 define('PARENT_MENU_ID', '32');
+define('HOW_TO_PARENT_MENU_ID', '23935');
 
 class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCycle {
 
@@ -83,32 +83,32 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
 
     public function addActionsAndFilters() {
 
-        if( $_GET['refresh_ibiza_menu'] == 1  ){
+        
+        if(is_admin() ){
+        
+            if ( isset( $_GET['refresh_ibiza_menu'] ) && $_GET['refresh_ibiza_menu'] == 1) {
 
-            add_action('init', array(&$this, 'do_menu'));
-            header( 'Location: /wp-admin/nav-menus.php' );
+                add_action('init', array(&$this, 'do_menu'));
+                header('Location: /wp-admin/nav-menus.php');
+            }
 
-        }
+            if ($_GET['promoted_menu'] == 1) {
 
-        if( $_GET['promoted_menu'] == 1  ){
+                add_action('init', array(&$this, 'promoted_menu'));
+            }
 
-            add_action('init', array(&$this, 'promoted_menu'));
-            
-        }
+            if ($_GET['promote_menu'] == 1 || $_GET['promote_menu'] == 2) {
 
-        if( $_GET['promote_menu'] == 1 || $_GET['promote_menu'] == 2 ){
+                add_action('init', array(&$this, 'promote_menu'));
+            }
 
-            add_action('init', array(&$this, 'promote_menu'));
-            
-        }
+            global $pagenow;
 
-        global $pagenow;
+            if ($pagenow == 'nav-menus.php') {
 
-        if( $pagenow == 'nav-menus.php' ){
-
-            wp_enqueue_script( 'menurefresh-js', get_template_directory_uri()  . '/assets/js/admin-scripts.js', array( 'jquery' ), '', true );
-            wp_enqueue_script( 'jquery.popupoverlay-js', get_template_directory_uri()  . '/vendor/jquery-popup-overlay/jquery.popupoverlay.js', array( 'jquery' ), '', true );
-
+                wp_enqueue_script('menurefresh-js', get_template_directory_uri() . '/assets/js/admin-scripts.js', array('jquery'), '', true);
+                wp_enqueue_script('jquery.popupoverlay-js', get_template_directory_uri() . '/vendor/jquery-popup-overlay/jquery.popupoverlay.js', array('jquery'), '', true);
+            }
         }
     }
 
@@ -117,10 +117,17 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
      * @return type
      */
     function get_json() {
-    global $ibiza_api;
-    
-        $jsonPath = $ibiza_api::api_location . '/ProductCatalog.Api/api/categorytree';  //get_template_directory() . '/assets/json/menu.json';
-        $cats = json_decode(file_get_contents($jsonPath));
+        global $ibiza_api;
+
+        $jsonPath = $ibiza_api::api . '/ProductCatalog.Api/categorytree';  //get_template_directory() . '/assets/json/menu.json';
+
+
+
+
+        $cats = json_decode(getSslPage($jsonPath));
+
+
+
         return $cats[0]->{TOP_NODE};
     }
 
@@ -134,26 +141,26 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
      * @param type $menu_type
      * @return type
      */
-    function update_menu($menu_id, $new_menu_id, $title, $menu_position = 0, $parent_menu_id = PARENT_MENU_ID, $menu_type = 'custom' , $page_link ='product-list' ,$db_id =0) {
+    function update_menu($menu_id, $new_menu_id, $title, $menu_position = 0, $parent_menu_id = PARENT_MENU_ID, $menu_type = 'custom', $page_link = 'product-list', $db_id = 0) {
 
         static $i = 0;
-        
-        if($page_link!='product-list'){
-            $i =1;
+
+        if ($page_link != 'product-list') {
+            $i = 1;
         }
-        
-        if($i ==1){
+
+        if ($i == 1) {
             $page_link = 'how-to-guides';
         }
-        
+
         return wp_update_nav_menu_item($menu_id, $db_id, array(
             'menu-item-parent-id' => $parent_menu_id,
-            'menu-item-db-id'   => $db_id , 
+            'menu-item-db-id' => $db_id,
             'menu-item-position' => $menu_position,
             'menu-item-object' => 'page',
             'menu-item-type' => $menu_type,
             'menu-item-status' => 'publish',
-            'menu-item-url' => '/'. $page_link .'/' . sanitize_title_with_dashes( $title ) . '/' . $new_menu_id ,
+            'menu-item-url' => '/' . $page_link . '/' . sanitize_title_with_dashes($title) . '/' . $new_menu_id,
             'menu-item-title' => $title,
         ));
     }
@@ -168,44 +175,44 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
 
         foreach ($data as $key1 => $dataIn) {
 
-            if($dataIn->title=='How To Guides'){
-                
-                update_post_meta(23935, 'cat-' . $dataIn->id, json_encode($dataIn) );
-                $parent_id = 23935;
+            if ($dataIn->title == 'How To Guides') {
+
+                update_post_meta(HOW_TO_PARENT_MENU_ID, 'cat-' . $dataIn->id, json_encode($dataIn));
+                $parent_id = HOW_TO_PARENT_MENU_ID;
                 if (count($dataIn->nodes) > 0) {
 
-                    $this->add_menu($dataIn->nodes, $menu_id, 23935);
-                }                 
-                
-            }else{
-                
-                
-                if($dataIn->title=='Shop'){
-                    update_post_meta(32, 'cat-' . $dataIn->id, json_encode($dataIn) );
+                    $this->add_menu($dataIn->nodes, $menu_id, HOW_TO_PARENT_MENU_ID);
+                }
+            } else {
+
+
+                if ($dataIn->title == 'Shop') {
+                    update_post_meta(PARENT_MENU_ID, 'cat-' . $dataIn->id, json_encode($dataIn));
                     $this->add_menu($dataIn->nodes, $menu_id, $child_menu_id);
-                }else{                
-                
-                
+                } else {
+
+
                     global $wpdb;
-                    
-                    if($parent_id == 23935){
+
+                    if ($parent_id == HOW_TO_PARENT_MENU_ID) {
                         $page_link = 'how-to-guides';
-                    }else{
+                    } else {
                         $page_link = 'product-list';
                     }
-                    
-                    $meta_key   = 'cat-' . $dataIn->id;
-                    $value      = $wpdb->get_var( $wpdb->prepare("SELECT post_id FROM $wpdb->postmeta WHERE meta_key = %s LIMIT 1" , $meta_key) );;                
-                    $db_id      = 0;
+
+                    $meta_key = 'cat-' . $dataIn->id;
+                    $value = $wpdb->get_var($wpdb->prepare("SELECT post_id FROM $wpdb->postmeta WHERE meta_key = %s LIMIT 1", $meta_key));
+                    ;
+                    $db_id = 0;
 
                     // check if menu is in our data
-                    if( isset($this->data[$value]) ){
+                    if (isset($this->data[$value])) {
                         $child_menu_id = $db_id = $value;
                         $order = $this->data[$value];
                         unset($this->data[$value]);
-                         $this->update_menu($menu_id, $dataIn->id, $dataIn->title,$order, $parent_id, 'custom' , $page_link,$db_id);
-                    }else{
-                        $child_menu_id = $this->update_menu($menu_id, $dataIn->id, $dataIn->title, $key1, $parent_id, 'custom' , $page_link,$db_id);
+                        $this->update_menu($menu_id, $dataIn->id, $dataIn->title, $order, $parent_id, 'custom', $page_link, $db_id);
+                    } else {
+                        $child_menu_id = $this->update_menu($menu_id, $dataIn->id, $dataIn->title, $key1, $parent_id, 'custom', $page_link, $db_id);
                     }
 
 
@@ -213,129 +220,117 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
 
 
 
-                    $dataIn->description = str_replace(array("\r\n", "\r", "\n"), "<br />", $dataIn->description); 
+                    $dataIn->description = str_replace(array("\r\n", "\r", "\n"), "<br />", $dataIn->description);
                     //update_post_meta($child_menu_id, 'cat-' . $dataIn->id, '1');
-                    update_post_meta($child_menu_id, 'cat-' . $dataIn->id, wp_json_encode($dataIn , JSON_UNESCAPED_UNICODE) );
+                    update_post_meta($child_menu_id, 'cat-' . $dataIn->id, wp_json_encode($dataIn, JSON_UNESCAPED_UNICODE));
                     // 1 has now use, ignore
                     if (count($dataIn->nodes) > 0) {
 
                         $this->add_menu($dataIn->nodes, $menu_id, $child_menu_id);
-                    }                
-                
+                    }
                 }
-                
             }
-            
-            
-
         }
     }
 
+    function promote_menu() {
 
-    function promote_menu()
-    {
-    
         $promote = '';
-        
-        if( $_GET['promoted'] == 1 ){
+
+        if ($_GET['promoted'] == 1) {
             $promote = 1;
         }
-        
+
         $col_to_update = 'post_content';
-        
-        if( $_GET['promote_menu'] == '2' ){
+
+        if ($_GET['promote_menu'] == '2') {
             $col_to_update = 'post_excerpt';
         }
-        
+
         // Update post 37
         $my_post = array(
-            'ID'           =>   (int)$_GET['menu_id'] ,
-            $col_to_update =>   $promote ,
+            'ID' => (int) $_GET['menu_id'],
+            $col_to_update => $promote,
         );
 
-      // Update the post into the database
-        wp_update_post( $my_post );        
-        
+        // Update the post into the database
+        wp_update_post($my_post);
     }
-    
-    
-    function promoted_menu()
-    {
-        
-        
-        $menu_item = get_post( $_GET['menu_id'] );
-        
+
+    function promoted_menu() {
+
+
+        $menu_item = get_post($_GET['menu_id']);
         ?>
         <div id="popupz">
-        <style>
+            <style>
 
-            .popup_content {
-                box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
-                display: none;
-                margin: 1em;
-            }
-            .popup_content {
-                background-color: #f5f5f5;
-                border: 1px solid #e3e3e3;
-                border-radius: 4px;
-                box-shadow: 0 1px 1px rgba(0, 0, 0, 0.05) inset;
-                margin-bottom: 20px;
-                min-height: 20px;
-                padding: 19px;
-            }
-        </style>
+                .popup_content {
+                    box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+                    display: none;
+                    margin: 1em;
+                }
+                .popup_content {
+                    background-color: #f5f5f5;
+                    border: 1px solid #e3e3e3;
+                    border-radius: 4px;
+                    box-shadow: 0 1px 1px rgba(0, 0, 0, 0.05) inset;
+                    margin-bottom: 20px;
+                    min-height: 20px;
+                    padding: 19px;
+                }
+            </style>
 
-        <div id="my_popup">
-           <!-- Add an optional button to close the popup -->
-           
-          
-           
-            <label>Promote category on category page</label>
-            <?php if( $menu_item->post_content == 1 ): ?>
-            <input id="promote_cat_cat" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" checked="checked" />
-            <?php else: ?>
-            <input id="promote_cat_cat" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" />
-            <?php endif; ?>     
-            <br />
-            <label>Promote category on landing page</label>
-            <?php if( $menu_item->post_excerpt == 1 ): ?>
-            <input id="promote_cat_landing" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" checked="checked" />
-            <?php else: ?>
-            <input id="promote_cat_landing" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" />
-            <?php endif; ?>
-            <button class="my_popup_close">Close</button>
+            <div id="my_popup">
+                <!-- Add an optional button to close the popup -->
 
-         </div>
-        <script>
-        jQuery(document).ready(function () {
-            
-            jQuery('#my_popup').popup( { autoopen : true , background : false , detach :true  ,  onclose: function() {
-                
-                
-                jQuery('#popupz,#my_popup_wrapper').remove();
-                
-            }} );
 
-        });    
-        </script>
+
+                <label>Promote category on category page</label>
+        <?php if ($menu_item->post_content == 1): ?>
+                    <input id="promote_cat_cat" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" checked="checked" />
+        <?php else: ?>
+                    <input id="promote_cat_cat" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" />
+        <?php endif; ?>     
+                <br />
+                <label>Promote category on landing page</label>
+        <?php if ($menu_item->post_excerpt == 1): ?>
+                    <input id="promote_cat_landing" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" checked="checked" />
+        <?php else: ?>
+                    <input id="promote_cat_landing" type="checkbox" value="<?php echo $menu_item->ID; ?>" name="" />
+        <?php endif; ?>
+                <button class="my_popup_close">Close</button>
+
+            </div>
+            <script>
+                jQuery(document).ready(function () {
+
+                    jQuery('#my_popup').popup({autoopen: true, background: false, detach: true, onclose: function () {
+
+
+                            jQuery('#popupz,#my_popup_wrapper').remove();
+
+                        }});
+
+                });
+            </script>
         </div>
-        <?php 
+        <?php
         die;
     }
-    
 
     /**
      * 
      */
     function do_menu() {
 
-        
-        
-        
-        
-        
-        
-        
+
+
+
+
+
+
+
 //                return wp_update_nav_menu_item('Main', 33959, array(
 //            'menu-item-parent-id' => 0,
 //            'menu-item-db-id'   => 33959 , 
@@ -346,12 +341,20 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
 //            'menu-item-url' => '#'  ,
 //            'menu-item-title' => 'updated',
 //        ));
-        
-        
-        
-        $jsonData       = $this->get_json();
 
-        $data           = array();
+
+
+        $jsonData = $this->get_json();
+
+        $data = array();
+
+
+
+       // ob_start();
+     //   print_r($cats);
+
+
+
 
         wp_nav_menu(array(
             'echo' => false,
@@ -361,34 +364,48 @@ class IbizaCategoriesPullPlugin_Plugin extends IbizaCategoriesPullPlugin_LifeCyc
         ));
 
 
-        $this->data     = $a->data;
-        $menuname       = 'Main';
-        $menu_exists    = wp_get_nav_menu_object($menuname);
-        $menu_id        = $menu_exists->term_id;
-        
+        $this->data = $a->data;
+        $menuname = 'Main';
+        $menu_exists = wp_get_nav_menu_object($menuname);
+        $menu_id = $menu_exists->term_id;
 
-       
+
+
         if ($menu_exists) {
             $this->add_menu($jsonData, $menu_id, $parent_id = PARENT_MENU_ID);
         }
-        
-        
-        
-        
+
+
+
+
 //        $stringData = print_r($this->data ,true);
 //     
 //        update_option( 'logs', $stringData, false );
 //        die;        
-        
-        foreach ($this->data as $key=> $data){
-            
-            wp_delete_post($key, true);
-            
+
+        if ($this->data) {
+
+            foreach ($this->data as $key => $data) {
+
+                wp_delete_post($key, true);
+            }
         }
         // delete any menus thats have been removed
-        
+
+
+
+//        $out2 = ob_get_contents();
+//
+//        ob_end_clean();
+//        $myFile = "testFiles.txt";
+//        $fh = fopen($myFile, 'w') or die("can't open file");
+//        $stringData = $out2;
+//        fwrite($fh, $stringData);
+//        fclose($fh);
     }
-    public $data ;
+
+    public $data;
+
 }
 
 /**
@@ -398,16 +415,15 @@ class Topbar1_Menu_Walker extends Walker_Nav_Menu {
 
     public $sub_menu_arr = array();
     public $data;
-    
-    
+
     function start_el(&$output, $item, $depth = 0, $args = array(), $id = 0) {
 
 
-        if ($item->menu_item_parent == 32 OR $item->menu_item_parent == 23935 OR in_array($item->menu_item_parent, $this->sub_menu_arr)) {
+        if ($item->menu_item_parent == PARENT_MENU_ID OR $item->menu_item_parent == HOW_TO_PARENT_MENU_ID OR in_array($item->menu_item_parent, $this->sub_menu_arr)) {
             $this->sub_menu_arr[$item->ID] = $item->ID;
-            
+
             $this->data[$item->ID] = $item->menu_order;
-            
+
             wp_delete_post($item->ID, true);
         }
     }
